use forwarded ip instead of real ip
The main purpose of this project is to provide a way to allow random
users to use ft without installing it. A bug was present in ft-demo
where every user got the same unique id. The way to fix this was
to use the IP provided from the header "X-Forwarded-For" instead of
Http.Request.RemoteAddr