do not parse forwarded ip
This fixes the issue with ids not being unique for each unique ip
address. Basically, main.go used some parsing function of net that parses
IP address. What this meant was that if an ip address was indeed invalid,
the net result would be the same for each IP address.